+++ bitcointx/core/__init__.py
@@ -24,6 +24,8 @@ from .serialize import (
     Hash, Hash160, make_mutable
 )
 
+from .util import _disable_boolean_use
+
 # Core definitions
 COIN = 100000000
 MAX_BLOCK_SIZE = 1000000
@@ -144,6 +146,7 @@ class _UintBitVector(ImmutableSerializable, metaclass=_UintBitVectorMeta):
                              .format(self._UINT_WIDTH_BYTES))
         object.__setattr__(self, 'data', data)
 
+    @_disable_boolean_use
     def is_null(self):
         return all(b == 0 for b in self.data)
 
@@ -203,6 +206,7 @@ class COutPoint(ImmutableSerializable):
         f.write(self.hash)
         f.write(struct.pack(b"<I", self.n))
 
+    @_disable_boolean_use
     def is_null(self):
         return ((self.hash == b'\x00'*32) and (self.n == 0xffffffff))
 
@@ -273,6 +277,7 @@ class CTxInBase(ImmutableSerializable):
         BytesSerializer.stream_serialize(self.scriptSig, f)
         f.write(struct.pack(b"<I", self.nSequence))
 
+    @_disable_boolean_use
     def is_final(self):
         return (self.nSequence == 0xffffffff)
 
@@ -338,6 +343,7 @@ class CBitcoinTxOut(CTxOutBase):
         f.write(struct.pack(b"<q", self.nValue))
         BytesSerializer.stream_serialize(self.scriptPubKey, f)
 
+    @_disable_boolean_use
     def is_valid(self):
         if not MoneyRange(self.nValue):
             return False
@@ -384,6 +390,7 @@ class CBitcoinTxInWitness(CTxInWitnessBase):
     def __init__(self, scriptWitness=script.CScriptWitness()):
         object.__setattr__(self, 'scriptWitness', scriptWitness)
 
+    @_disable_boolean_use
     def is_null(self):
         return self.scriptWitness.is_null()
 
@@ -442,6 +449,7 @@ class CBitcoinTxWitness(CTxWitnessBase):
                                  else self._txin_witness_class.from_txin_witness(w)
                                  for w in vtxinwit))
 
+    @_disable_boolean_use
     def is_null(self):
         for n in range(len(self.vtxinwit)):
             if not self.vtxinwit[n].is_null():
@@ -524,6 +532,7 @@ class CTransactionBase(ImmutableSerializable, ReprOrStrMixin):
         object.__setattr__(self, 'vout', tuple(self._txout_class.from_txout(txout) for txout in vout))
         object.__setattr__(self, 'wit', self._witness_class.from_witness(witness))
 
+    @_disable_boolean_use
     def is_coinbase(self):
         return len(self.vin) == 1 and self.vin[0].prevout.is_null()