+++ bitcointx/core/script.py
@@ -28,6 +28,8 @@ import bitcointx.core._bignum
 
 from .serialize import VarIntSerializer, BytesSerializer, ImmutableSerializable
 
+from .util import _disable_boolean_use
+
 _script_class_params = {}  # to be filled by _SetScriptClassParams()
 _raw_signature_hash_func = None  # to be set by _SetScriptClassParams()
 
@@ -82,6 +84,7 @@ class CScriptOp(int):
 
         return int(self - OP_1+1)
 
+    @_disable_boolean_use
     def is_small_int(self):
         """Return true if the op pushes a small integer to the stack"""
         if 0x51 <= self <= 0x60 or self == 0:
@@ -674,6 +677,7 @@ class CScriptBase(bytes):
 
         return "CScript([%s])" % ', '.join(ops)
 
+    @_disable_boolean_use
     def is_p2sh(self):
         """Test if the script is a p2sh scriptPubKey
 
@@ -684,6 +688,7 @@ class CScriptBase(bytes):
                 self[1] == 0x14 and
                 self[22] == OP_EQUAL)
 
+    @_disable_boolean_use
     def is_witness_scriptpubkey(self):
         """Returns true if this is a scriptpubkey signaling segregated witness data.
 
@@ -711,22 +716,27 @@ class CScriptBase(bytes):
         """Returns the witness program"""
         return self[2:]
 
+    @_disable_boolean_use
     def is_witness_v0_keyhash(self):
         """Returns true if this is a scriptpubkey for V0 P2WPKH. """
         return len(self) == 22 and self[0:2] == b'\x00\x14'
 
+    @_disable_boolean_use
     def is_witness_v0_nested_keyhash(self):
         """Returns true if this is a scriptSig for V0 P2WPKH embedded in P2SH. """
         return len(self) == 23 and self[0:3] == b'\x16\x00\x14'
 
+    @_disable_boolean_use
     def is_witness_v0_scripthash(self):
         """Returns true if this is a scriptpubkey for V0 P2WSH. """
         return len(self) == 34 and self[0:2] == b'\x00\x20'
 
+    @_disable_boolean_use
     def is_witness_v0_nested_scripthash(self):
         """Returns true if this is a scriptSig for V0 P2WSH embedded in P2SH. """
         return len(self) == 35 and self[0:3] == b'\x22\x00\x20'
 
+    @_disable_boolean_use
     def is_push_only(self):
         """Test if the script only contains pushdata ops
 
@@ -775,10 +785,12 @@ class CScriptBase(bytes):
             return False
         return True
 
+    @_disable_boolean_use
     def is_unspendable(self):
         """Test if the script is provably unspendable"""
         return (len(self) > 0 and self[0] == OP_RETURN) or len(self) > MAX_SCRIPT_SIZE
 
+    @_disable_boolean_use
     def is_valid(self):
         """Return True if the script is valid, False otherwise
 
@@ -821,6 +833,7 @@ class CScriptBase(bytes):
 
         return (genesis_hash, pegout_scriptpubkey)
 
+    @_disable_boolean_use
     def is_pegout(self):
         return self.get_pegout_data() is not None
 
@@ -879,6 +892,7 @@ class CScriptWitness(ImmutableSerializable):
     def __repr__(self):
         return 'CScriptWitness([' + ','.join("x('%s')" % bitcointx.core.b2x(s) for s in self.stack) + '])'
 
+    @_disable_boolean_use
     def is_null(self):
         return len(self.stack) == 0